fixes initial window size to picture with retina scaling
The scaling of retina screens lets the window be bigger than the actual
picture displayed due to the "downscaling" of the screen to the window
manager. The `- BackingScaleFactor` of NSWindow states what scaling is
used for the screen.
This applied to the image size will fit the window to the size of the
image itself.

Signed-off-by: Jakob Wedemeyer <jakob@soerensen.com>